+++ client/test/pages/createRequestTest.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { expect as e } from 'chai';
 import { shallow } from 'enzyme';
-import { CreateRequest } from '../../pages/createrequest';
+import { CreateRequest } from '../../pages/CreateRequest';
 
 describe('<CreateRequest />', () => {
   let wrapper;
